Apply more suitable types

+++ src/telegram/callbackgame.h
@@ -9,12 +9,12 @@ struct CallbackGame
 {
     typedef QSharedPointer<CallbackGame> Ptr;
 
-    int32_t m_user_id              = 0;
-    int32_t m_score                = 0;
+    qint64  m_user_id              = 0;
+    qint64  m_score                = 0;
     bool    m_force                = false;
     bool    m_disable_edit_message = false;
-    int32_t m_chat_id              = 0;
-    int32_t m_message_id           = 0;
+    qint64  m_chat_id              = 0;
+    qint64  m_message_id           = 0;
     QString m_inline_message_id;
 };
